add new in answer quality

+++ 2024FallCS209A-ProjectDemo-main/2024FallCS209A-ProjectDemo-main/src/main/java/CS209A/project/demo/controller/AnswerController.java
@@ -4,7 +4,6 @@ package CS209A.project.demo.controller;
 
 import CS209A.project.demo.service.AnswerService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Map;
@@ -40,7 +39,7 @@ public class AnswerController {
         return "Hello from another endpoint!";
     }
     @GetMapping("/api/answer-quality")
-    public Map<String, Map<String, Long>> getAnswerQualityData() {
+    public Map<String, Map<String, Double>> getAnswerQualityData() {
         return answerService.getAllImpactData();
     }
